+++ oshdb-api/src/test/java/org/heigit/bigspatialdata/oshdb/api/tests/TestOSMContributionGetContributorUserId.java
@@ -12,6 +12,7 @@ import java.util.Collections;
 import java.util.EnumSet;
 import java.util.List;
 import org.heigit.bigspatialdata.oshdb.api.object.OSMContribution;
+import org.heigit.bigspatialdata.oshdb.osh.OSHEntity;
 import org.heigit.bigspatialdata.oshdb.osh.OSHNode;
 import org.heigit.bigspatialdata.oshdb.osh.OSHWay;
 import org.heigit.bigspatialdata.oshdb.osm.OSMMember;
@@ -32,12 +33,17 @@ public class TestOSMContributionGetContributorUserId {
   public TestOSMContributionGetContributorUserId() throws Exception {
   }
 
+  private final OSHEntity dummyOshEntity = OSHNode.build(Collections.singletonList(
+      new OSMNode(-1L, 1, new OSHDBTimestamp(0L), 1L, 1, new int[]{}, 0, 0)
+  ));
+
   @Test
   public void node() throws Exception {
     // timestamp match
     OSMContribution c = new OSMContribution(new IterateAllEntry(
         new OSHDBTimestamp(123),
-        new OSMNode(1L, 1, new OSHDBTimestamp(123L), 1L, 7, new int[] {}, 0, 0), null,
+        new OSMNode(1L, 1, new OSHDBTimestamp(123L), 1L, 7, new int[]{}, 0, 0), null,
+        dummyOshEntity,
         null, null, null, null,
         new LazyEvaluatedContributionTypes(EnumSet.of(ContributionType.CREATION)),
         1L
@@ -47,6 +53,7 @@ public class TestOSMContributionGetContributorUserId {
     c = new OSMContribution(new IterateAllEntry(
         new OSHDBTimestamp(123),
         new OSMNode(1L, 1, new OSHDBTimestamp(122L), 1L, 7, new int[] {}, 0, 0), null,
+        dummyOshEntity,
         null, null, null, null,
         new LazyEvaluatedContributionTypes(EnumSet.of(ContributionType.CREATION)),
         1L
@@ -56,6 +63,7 @@ public class TestOSMContributionGetContributorUserId {
         new OSHDBTimestamp(123),
         new OSMNode(1L, 2, new OSHDBTimestamp(122L), 2L, 7, new int[] {3,4}, 0, 0),
         new OSMNode(1L, 1, new OSHDBTimestamp(121L), 1L, 6, new int[] {1,2}, 0, 0),
+        dummyOshEntity,
         null, null, null, null,
         new LazyEvaluatedContributionTypes(EnumSet.of(ContributionType.TAG_CHANGE)),
         2L
@@ -65,6 +73,7 @@ public class TestOSMContributionGetContributorUserId {
         new OSHDBTimestamp(123),
         new OSMNode(1L, -2, new OSHDBTimestamp(122L), 2L, 7, new int[] {}, 0, 0), // negative version == isVisible = false
         new OSMNode(1L, 1, new OSHDBTimestamp(121L), 1L, 6, new int[] {}, 0, 0),
+        dummyOshEntity,
         null, null, null, null,
         new LazyEvaluatedContributionTypes(EnumSet.of(ContributionType.DELETION)),
         2L
@@ -75,6 +84,7 @@ public class TestOSMContributionGetContributorUserId {
         new OSHDBTimestamp(123),
         new OSMNode(1L, 1, new OSHDBTimestamp(122L), 1L, 7, new int[] {}, 0, 0),
         new OSMNode(1L, 1, new OSHDBTimestamp(122L), 1L, 7, new int[] {}, 0, 0),
+        dummyOshEntity,
         null, null, null, null,
         new LazyEvaluatedContributionTypes(EnumSet.noneOf(ContributionType.class)),
         -1L
@@ -87,6 +97,7 @@ public class TestOSMContributionGetContributorUserId {
     OSMContribution c = new OSMContribution(new IterateAllEntry(
         new OSHDBTimestamp(123),
         new OSMWay(1L, 1, new OSHDBTimestamp(123L), 1L, 7, new int[] {}, new OSMMember[] {}), null,
+        dummyOshEntity,
         null, null, null, null,
         new LazyEvaluatedContributionTypes(EnumSet.of(ContributionType.CREATION)),
         1L
@@ -107,6 +118,7 @@ public class TestOSMContributionGetContributorUserId {
     OSMContribution c = new OSMContribution(new IterateAllEntry(
         new OSHDBTimestamp(123),
         entity, entity,
+        dummyOshEntity,
         null, null, null, null,
         new LazyEvaluatedContributionTypes(EnumSet.of(ContributionType.GEOMETRY_CHANGE)),
         1L
@@ -119,6 +131,7 @@ public class TestOSMContributionGetContributorUserId {
     OSMContribution c = new OSMContribution(new IterateAllEntry(
         new OSHDBTimestamp(123),
         new OSMRelation(1L, 1, new OSHDBTimestamp(123L), 1L, 7, new int[] {}, new OSMMember[] {}), null,
+        dummyOshEntity,
         null, null, null, null,
         new LazyEvaluatedContributionTypes(EnumSet.of(ContributionType.CREATION)),
         1L
@@ -139,6 +152,7 @@ public class TestOSMContributionGetContributorUserId {
     OSMContribution c = new OSMContribution(new IterateAllEntry(
         new OSHDBTimestamp(123),
         entity, entity,
+        dummyOshEntity,
         null, null, null, null,
         new LazyEvaluatedContributionTypes(EnumSet.of(ContributionType.GEOMETRY_CHANGE)),
         1L
@@ -164,6 +178,7 @@ public class TestOSMContributionGetContributorUserId {
     OSMContribution c = new OSMContribution(new IterateAllEntry(
         new OSHDBTimestamp(123),
         entity, entity,
+        dummyOshEntity,
         null, null, null, null,
         new LazyEvaluatedContributionTypes(EnumSet.of(ContributionType.GEOMETRY_CHANGE)),
         1L